backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
scipy
scikit-learn
statsmodels
openpyxl
anthropic
python-dotenv
sqlalchemy
aiofiles
